requirements.txt
amqp==5.1.1
asgiref==3.4.1
beautifulsoup4==4.10.0
billiard==3.6.4.0
celery==5.2.7
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.9
django-admin-json-editor==0.2.3
django-allauth==0.45.0
django-appconf==1.0.5
django-bootstrap4==21.2
django-braintree==1.3.4
django-celery-email==3.0.0
django-ckeditor==6.2.0
django-grappelli==3.0.3
django-js-asset==1.2.2
django-json-widget==1.1.1
django-mptt==0.13.4
django-spurl==0.6.8
dnspython==2.2.1
et-xmlfile==1.1.0
eventlet==0.33.1
flower==1.1.0
future==0.18.2
greenlet==1.1.2
humanize==4.2.3
idna==3.3
jsonfield==3.1.0
kombu==5.2.4
lml==0.1.0
oauthlib==3.1.1
Pillow==8.4.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
psycopg2-binary==2.9.2
pycparser==2.20
PyJWT==2.3.0
python-gettext==4.0
python3-openid==3.2.0
pytz==2022.1
rabbitmq==0.2.0
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
standardjson==0.3.1
texttable==1.6.4
tornado==6.2
urllib3==1.26.7
URLObject==2.4.3
vine==5.0.0
waitress==2.1.2
wcwidth==0.2.5
whitenoise==5.3.0
